+++ stone-js-cloud-file/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.8.3
+
+### Patch Changes
+
+- @stone-js/core@0.8.3
+- @stone-js/config@0.8.3
+- @stone-js/filesystem@0.8.3
+
 ## 0.8.2
 
 ### Patch Changes