+++ stone-js-filesystem/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.8.3
+
+### Patch Changes
+
+- @stone-js/core@0.8.3
+
 ## 0.8.2
 
 ### Patch Changes